requirements.txt
altair==5.3.0
annotated-types==0.6.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dotenv==0.0.5
filelock==3.14.0
flake8==7.0.0
Flask==3.0.3
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pyflakes==3.2.0
pygame==2.5.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
six==1.16.0
smmap==5.0.1
streamlit==1.34.0
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vboxapi==1.0
watchdog==4.0.0
Werkzeug==3.0.3
